feat: create a `types` module
