+++ src/chains/arbitrum/index.ts
@@ -1,5 +1,5 @@
 import { arbitrum as arbitrumMetadata } from '@wagmi/chains';
-import { Chain } from '@/chains';
+import { Chain } from '@/types/chain';
 import { opcodes } from './vm/opcodes';
 import { precompiles } from './vm/precompiles';
