+++ src/chains/mainnet/vm/opcodes/arithmetic/index.ts
@@ -1,4 +1,4 @@
-import { Opcode } from '@/chains/types';
+import { Opcode } from '@/types/opcode';
 import { add } from './add';
 
 export const opcodes: Opcode[] = [add];
