+++ src/chains/mainnet/vm/opcodes/index.ts
@@ -1,4 +1,4 @@
-import { Opcode } from '@/chains/types';
+import { Opcode } from '@/types/opcode';
 import { opcodes as arithmeticOpcodes } from './arithmetic';
 import { opcodes as blockOpcodes } from './block';
 import { opcodes as memoryOpcodes } from './memory';
