+++ src/chains/mainnet/vm/opcodes/memory/mstore.ts
@@ -1,4 +1,4 @@
-import { Opcode } from '@/chains/types';
+import { Opcode } from '@/types/opcode';
 
 export const mstore: Opcode = {
   number: 52,
