+++ src/chains/optimism/vm/opcodes/block/number.ts
@@ -1,4 +1,4 @@
-import { Opcode } from '@/chains/types';
+import { Opcode } from '@/types/opcode';
 
 export const number: Opcode = {
   number: 43,
