+++ src/chains/optimism/vm/opcodes/index.ts
@@ -1,6 +1,6 @@
 import { opcodes as mainnetArithmeticOpcodes } from '@/chains/mainnet/vm/opcodes/arithmetic';
 import { opcodes as mainnetMemoryOpcodes } from '@/chains/mainnet/vm/opcodes/memory';
-import { Opcode } from '@/chains/types';
+import { Opcode } from '@/types/opcode';
 import { opcodes as blockOpcodes } from './block';
 
 const modifiedOpcodes: Opcode[] = [...blockOpcodes];
