+++ src/components/ui/ChainDiffSelectorChainCombobox.tsx
@@ -2,8 +2,8 @@ import { useEffect, useState } from 'react';
 import Image from 'next/image';
 import { Combobox } from '@headlessui/react';
 import { CheckIcon, ChevronUpDownIcon } from '@heroicons/react/20/solid';
-import { Chain } from '@/chains/types';
 import { classNames } from '@/lib/utils';
+import { Chain } from '@/types/chain';
 
 interface Props {
   label: string;
