+++ src/types/opcode.ts
@@ -1,13 +1,3 @@
-import { Chain as Metadata } from '@wagmi/chains';
-import { Address } from 'viem';
-
-type PrecompileParam = {
-  byteStart: number;
-  byteLength: number;
-  name: string;
-  description: string;
-};
-
 type Variable = {
   name: string;
   description: string;
@@ -32,23 +22,6 @@ type Example = {
   memory?: Memory;
 };
 
-export type Predeploy = {
-  address: Address;
-  name: string;
-  description: string;
-  deprecated: boolean;
-};
-
-export type Precompile = {
-  address: Address;
-  name: string;
-  description: string;
-  minGas: number;
-  input: PrecompileParam[];
-  output: PrecompileParam[];
-  references: string[];
-};
-
 export type Opcode = {
   number: number;
   name: string;
@@ -62,10 +35,4 @@ export type Opcode = {
   errorCases: string[];
   notes: string[];
   references: string[];
-};
-
-export type Chain = {
-  metadata: Metadata;
-  precompiles: (Precompile | Predeploy)[];
-  opcodes: Opcode[];
 };
